+++ gems/aws-sdk-sqs/spec/client/verify_checksums_spec.rb
@@ -31,6 +31,25 @@ module Aws
           }
         }}
 
+        let(:message_system_attributes) {{
+          'ccc' => {
+            string_value: 'test',
+            data_type: 'String'
+          },
+          aaa: {
+            binary_value: [ 2, 3, 4 ].pack('C*'),
+            data_type: 'Binary'
+          },
+          zzz: {
+            data_type: 'Number',
+            string_value: '0230.01'
+          },
+          'öther_encodings' => {
+            data_type: 'String',
+            string_value: 'Tüst'.encode('ISO-8859-1')
+          }
+        }}
+
         it 'is enabled by default' do
           client = Client.new(
             region: 'us-east-1',
@@ -52,13 +71,15 @@ module Aws
 
         describe '#send_message' do
           let(:md5_of_attributes) { '756d7f4338696745d063b420a2f7e502' }
+          let(:md5_of_system_attributes) { '756d7f4338696745d063b420a2f7e502' }
 
           before(:each) do
             response_body = <<-JSON
               {
-               "MD5OfMessageAttributes": "#{md5_of_attributes}",
-               "MD5OfMessageBody": "900150983cd24fb0d6963f7d28e17f72",
-               "MessageId": "5fea7756-0ea4-451a-a703-a558b933e274"
+                "MessageId": "5fea7756-0ea4-451a-a703-a558b933e274",
+                "MD5OfMessageBody": "900150983cd24fb0d6963f7d28e17f72",
+                "MD5OfMessageAttributes": "#{md5_of_attributes}",
+                "MD5OfMessageSystemAttributes": "#{md5_of_system_attributes}" 
               }
             JSON
 
@@ -77,6 +98,7 @@ module Aws
                 queue_url:'https://queue.url',
                 message_body: message_body,
                 message_attributes: message_attributes,
+                message_system_attributes: message_system_attributes
               )
             }.not_to raise_error
           end
@@ -87,16 +109,21 @@ module Aws
                 queue_url:'https://queue.url',
                 message_body: message_body,
                 message_attributes: {},
+                message_system_attributes: {}
               )
             }.not_to raise_error
           end
 
           context 'when data types have custom labels' do
             let(:md5_of_attributes) { '5b7ef6c8a8d46001c7cdadaeea917aa4' }
+            let(:md5_of_system_attributes) { '5b7ef6c8a8d46001c7cdadaeea917aa4' }
 
             before(:each) do
               message_attributes.keys.each do |attribute_name|
                 message_attributes[attribute_name][:data_type] += '.test'
+              end
+              message_system_attributes.keys.each do |attribute_name|
+                message_system_attributes[attribute_name][:data_type] += '.test'
               end
             end
 
@@ -106,6 +133,7 @@ module Aws
                   queue_url:'https://queue.url',
                   message_body: message_body,
                   message_attributes: message_attributes,
+                  message_system_attributes: message_system_attributes
                 )
               }.not_to raise_error
             end
@@ -117,17 +145,20 @@ module Aws
                 queue_url:'https://queue.url',
                 message_body: message_body + 'junk',
                 message_attributes: message_attributes,
+                message_system_attributes: message_system_attributes
               )
             }.to raise_error(Aws::Errors::ChecksumError)
           end
 
           it 'raises when the md5 checksums do not match for the body' do
             message_attributes['ccc'][:string_value] += 'junk'
+            message_system_attributes['ccc'][:string_value] += 'junk'
             expect {
               client.send_message(
                 queue_url:'https://queue.url',
                 message_body: message_body,
                 message_attributes: message_attributes,
+                message_system_attributes: message_system_attributes
               )
             }.to raise_error(Aws::Errors::ChecksumError)
           end
@@ -146,6 +177,7 @@ module Aws
                 queue_url:'https://queue.url',
                 message_body: message_body,
                 message_attributes: message_attributes,
+                message_system_attributes: message_system_attributes
               )
             }.to raise_error(Aws::SQS::Errors::ServiceError)
           end
@@ -153,7 +185,6 @@ module Aws
         end
 
         describe '#send_message_batch' do
-
           before(:each) do
             client.handle(step: :send) do |context|
               context.http_response.signal_done(
@@ -165,7 +196,8 @@ module Aws
                     {
                       "Id": "msg-id",
                       "MD5OfMessageBody": "900150983cd24fb0d6963f7d28e17f72",
-                      "MD5OfMessageAttributes": "756d7f4338696745d063b420a2f7e502"
+                      "MD5OfMessageAttributes": "756d7f4338696745d063b420a2f7e502",
+                      "MD5OfMessageSystemAttributes": "756d7f4338696745d063b420a2f7e502"
                     }
                   ]
                 }
@@ -183,6 +215,7 @@ module Aws
                     id: 'msg-id',
                     message_body: message_body,
                     message_attributes: message_attributes,
+                    message_system_attributes: message_system_attributes
                   }
                 ]
               )
@@ -198,6 +231,7 @@ module Aws
                     id: 'msg-id',
                     message_body: message_body + 'junk',
                     message_attributes: message_attributes,
+                    message_system_attributes: message_system_attributes
                   }
                 ]
               )
@@ -214,6 +248,7 @@ module Aws
                     id: 'msg-id',
                     message_body: message_body,
                     message_attributes: message_attributes,
+                    message_system_attributes: message_system_attributes
                   }
                 ]
               )
@@ -245,6 +280,7 @@ module Aws
                     id: 'msg-id',
                     message_body: message_body,
                     message_attributes: message_attributes,
+                    message_system_attributes: message_system_attributes
                   }
                 ]
               )